pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = "flask app"
authors = ["ao <noemails@to.me>"]
readme = "README.md"
packages = [{include = "page_analyzer"}]

[tool.poetry.dependencies]
python = "^3.10"
flask = "^2.3.2"
python-dotenv = "^1.0.0"
psycopg2-binary = "^2.9.5"
validators = "0.20.0"
beautifulsoup4 = "^4.12.2"
# gunicorn = "^21.2.0"
# psycopg2 = "^2.9.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
